remote add: use strvec to store tracking branches
Store the list of branches to track in a ’struct strvec' instead of a
'struct string_list'. This in preparation for the next commit where it
will be convenient to have them stored in a NULL terminated array. This
means that we now duplicate the strings when storing them but the
overhead is not significant.

Signed-off-by: Phillip Wood <phillip.wood@dunelm.org.uk>