t3404: avoid losing exit status with focus on `git show` and `git cat-files`
The exit code of the preceding command in a pipe is disregarded. So
if that preceding command is a Git command that fails, the test would
not fail. Instead, by saving the output of that Git command to a file,
and removing the pipe, we make sure the test will fail if that Git
command fails. This particular patch focuses on some of the instances
which include `git show` and `git cat-files`.

Signed-off-by: Usman Akinyemi <usmanakinyemi202@gmail.com>